rename recipe list
